% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{get_all_anomalies}
\alias{get_all_anomalies}
\title{Полное обнаружение всех видов аномалий (с поддержкой периода)}
\usage{
get_all_anomalies(
  conn,
  username = NULL,
  limit = 1000,
  since = NULL,
  until = NULL
)
}
\description{
Полное обнаружение всех видов аномалий (с поддержкой периода)
}
