% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{get_anomalies_for_developer}
\alias{get_anomalies_for_developer}
\title{Аномалии для конкретного разработчика}
\usage{
get_anomalies_for_developer(conn, username)
}
\arguments{
\item{conn}{Подключение к БД}

\item{username}{Никнейм разработчика}
}
\description{
Аномалии для конкретного разработчика
}
