% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_change_volume}
\alias{get_change_volume}
\title{Анализ объема изменений}
\usage{
get_change_volume(conn, username = NULL, email = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}

\item{email}{Email разработчика (опционально)}
}
\value{
data.frame с результатами анализа
}
\description{
Анализ объема изменений
}
