% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_commit_size_analysis}
\alias{get_commit_size_analysis}
\title{Анализ размера коммитов (классификация)}
\usage{
get_commit_size_analysis(conn, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с классификацией коммитов
}
\description{
Анализ размера коммитов (классификация)
}
