% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_developer_profile_by_email}
\alias{get_developer_profile_by_email}
\title{Полный профиль разработчика (по email)}
\usage{
get_developer_profile_by_email(conn, email)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{email}{Email разработчика (обязательный)}
}
\value{
list с результатами анализа
}
\description{
Полный профиль разработчика (по email)
}
