% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_developer_profile_by_username}
\alias{get_developer_profile_by_username}
\title{Полный профиль разработчика (по нику)}
\usage{
get_developer_profile_by_username(conn, username)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (обязательный)}
}
\value{
list с результатами анализа
}
\description{
Полный профиль разработчика (по нику)
}
