% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_file_change_frequency}
\alias{get_file_change_frequency}
\title{Анализ работы с конкретными файлами/директориями}
\usage{
get_file_change_frequency(conn, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с частотой изменений файлов
}
\description{
Анализ работы с конкретными файлами/директориями
}
