% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_hourly_distribution}
\alias{get_hourly_distribution}
\title{Часовое распределение с процентами}
\usage{
get_hourly_distribution(conn, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с результатами анализа
}
\description{
Часовое распределение с процентами
}
