% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_preferred_languages}
\alias{get_preferred_languages}
\title{Предпочитаемые языки программирования}
\usage{
get_preferred_languages(conn, email = NULL, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{email}{Email разработчика (опционально)}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с статистикой по расширениям файлов
}
\description{
Предпочитаемые языки программирования
}
