% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_productivity_daily}
\alias{get_productivity_daily}
\title{Анализ продуктивности (коммиты и изменения по дням)}
\usage{
get_productivity_daily(conn, email = NULL, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{email}{Email разработчика (опционально)}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с ежедневной статистикой
}
\description{
Анализ продуктивности (коммиты и изменения по дням)
}
