% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_sensitive_files_analysis}
\alias{get_sensitive_files_analysis}
\title{Работа с чувствительными файлами (безопасность)}
\usage{
get_sensitive_files_analysis(conn, email = NULL, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{email}{Email разработчика (опционально)}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с анализом чувствительных изменений
}
\description{
Работа с чувствительными файлами (безопасность)
}
