% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Сводная статистика по всем разработчикам}
\usage{
get_summary_stats(conn)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}
}
\value{
data.frame со сводной статистикой
}
\description{
Сводная статистика по всем разработчикам
}
