% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_team_metrics}
\alias{get_team_metrics}
\title{Получить метрики команды (можно фильтровать по списку имён)}
\usage{
get_team_metrics(conn, usernames = NULL)
}
\description{
Получить метрики команды (можно фильтровать по списку имён)
}
