% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_time_patterns}
\alias{get_time_patterns}
\title{Анализ временных паттернов по часам}
\usage{
get_time_patterns(conn, username = NULL, email = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}

\item{email}{Email разработчика (опционально)}
}
\value{
data.frame с результатами анализа
}
\description{
Анализ временных паттернов по часам
}
