% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{get_top_anomaly_developers}
\alias{get_top_anomaly_developers}
\title{Топ разработчиков по аномалиям}
\usage{
get_top_anomaly_developers(anomalies, n = 5)
}
\description{
Топ разработчиков по аномалиям
}
