% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_weekday_patterns}
\alias{get_weekday_patterns}
\title{Анализ активности по дням недели}
\usage{
get_weekday_patterns(conn, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с результатами анализа
}
\description{
Анализ активности по дням недели
}
