\name{GitTest-package}
\alias{GitTest-package}
\alias{GitTest}
\docType{package}
\title{
\packageTitle{GitTest}
}
\description{
\packageDescription{GitTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GitTest}
En este paquete hay una serie de funciones prueba: rest, es.divisible, IMC y IMG
\packageIndices{GitTest}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{Ramón Nieto Domínguez
\packageAuthor{GitTest}

Maintainer: \packageMaintainer{GitTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
