% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{parse_commit}
\alias{parse_commit}
\title{Parse a single commit block into a data frame of changes}
\usage{
parse_commit(block, repo_id)
}
\arguments{
\item{block}{Character vector representing one commit (from get_commits())}

\item{repo_id}{Numeric ID of the repository}
}
\value{
Data frame with columns: commit, src_file, dst_file, start_del, count_del, start_add, count_add, added_code, deleted_code
}
\description{
Parse a single commit block into a data frame of changes
}
