% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{get_all_anomalies}
\alias{get_all_anomalies}
\title{Обнаружение аномалий в Git-активности
Полное обнаружение всех видов аномалий}
\usage{
get_all_anomalies(conn)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}
}
\value{
data.frame со всеми аномалиями
}
\description{
Обнаружение аномалий в Git-активности
Полное обнаружение всех видов аномалий
}
