% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_most_productive_days}
\alias{get_most_productive_days}
\title{Самые продуктивные дни}
\usage{
get_most_productive_days(conn, email = NULL, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{email}{Email разработчика (опционально)}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с топ днями
}
\description{
Самые продуктивные дни
}
