% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{get_work_breaks}
\alias{get_work_breaks}
\title{Анализ перерывов в работе}
\usage{
get_work_breaks(conn, username = NULL)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{username}{Никнейм разработчика (опционально)}
}
\value{
data.frame с анализом перерывов
}
\description{
Анализ перерывов в работе
}
