% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_developers.R
\name{search_developers}
\alias{search_developers}
\title{Поиск разработчиков по имени или email}
\usage{
search_developers(conn, search_term)
}
\arguments{
\item{conn}{Подключение к базе данных DuckDB}

\item{search_term}{Поисковый запрос}
}
\value{
data.frame с найденными разработчиками
}
\description{
Поиск разработчиков по имени или email
}
